interleaving: clarify types in bit_array_interleave
Follow-up 76e922e441f2639f0c8f3a231b80d75a48684427

+++ bit_array_interleave.h
@@ -6,7 +6,11 @@
 struct bit_array_interleave_lookup_table {
 	/**
 	 * Contains octet lookup table with
-	 * shifts for each dimension
+	 * shifts for each dimension.
+	 * Structure:
+	 *   - dimensions: [1; 20]
+	 *   - lookup table 2^8 values: [0; 255]
+	 *   - bit array
 	 */
 	bit_array ***tables;
 	/**
@@ -16,11 +20,11 @@ struct bit_array_interleave_lookup_table {
 	/**
 	 * Amount of dimensions
 	 */
-	size_t dim;
+	uint8_t dim;
 };
 
 struct bit_array_interleave_lookup_table *
-bit_array_interleave_new_lookup_tables(size_t dim);
+bit_array_interleave_new_lookup_tables(uint8_t dim);
 
 void
 bit_array_interleave_free_lookup_tables(
